Feature/369: Shorter Import Paths Using Absolute TypeScript Paths (#29)
* Fix directory structure for domain according to conventions

* Modify config files for absolute paths

* Add module path mapper in jest BDD config

* Ignore audit issues for now

* Modify config files

* Add ESLint rule for default import

* Fix Logger call for thirdparty request object

* Replace all unit test import paths with short absolute paths

* Fix import path in consents handler

* Fix compile error in importing config in test folder

* Modify npm start to use tsconfig.build.json

* Modify jest coverage config to resolve import paths

* Remove trailing commas in eslintrc config

* fix: correct overwritten jest type for an integration test

* fix: fix all lint and module not found errors

* fix: npm start script tsconfig parameters

* refactor: change handler imports to shorter absolute paths

* refactor: change test imports to shorter absolute paths

* fix: ignore huge audit file using production flag
