+++ jest.coverage.config.js
@@ -1,5 +1,8 @@
 'use strict'
 
+const { pathsToModuleNameMapper } = require('ts-jest/utils')
+const { compilerOptions } = require('./tsconfig')
+
 module.exports = {
   verbose: true,
   preset: 'ts-jest',
@@ -17,5 +20,8 @@ module.exports = {
       lines: 90
     }
   },
+  moduleNameMapper: pathsToModuleNameMapper(compilerOptions.paths, {
+    prefix: '<rootDir>/'
+  }),
   reporters: ['jest-junit', 'default']
 }
