+++ package.json
@@ -11,18 +11,18 @@
     "knex": "--knexfile ./config/knexfile.ts"
   },
   "scripts": {
-    "audit:resolve": "SHELL=sh resolve-audit",
-    "audit:check": "SHELL=sh check-audit",
-    "build": "tsc",
+    "audit:resolve": "SHELL=sh resolve-audit --production",
+    "audit:check": "SHELL=sh check-audit --production",
+    "build": "tsc -p ./tsconfig.build.json",
     "docker:build": "docker build -t auth-service:local -f ./Dockerfile ./",
     "docker:run": "docker run -p 4004:4004 auth-service:local",
     "lint": "eslint ./src/**/*.ts *.js",
     "pretest": "echo \"pretest - TODO...\"",
     "release": "standard-version",
     "standard": "echo '\\033[1;33m  This project uses eslint instead of standard. Use `npm run lint` instead.'",
-    "start": "ts-node -P ./tsconfig.json ./src/cli.ts",
+    "start": "ts-node -r tsconfig-paths/register -P ./tsconfig.build.json ./src/cli.ts",
     "dev": "nodemon --watch 'src/**/*.ts' --exec 'ts-node' src/cli.ts",
-    "test": "npm run test:unit",
+    "test": "npm run test:unit test:integration",
     "test:bdd": "jest --config './jest.bdd.config.js' --testMatch '**/test/step-definitions/**/*.step.ts'",
     "test:coverage": "jest --coverage --coverageThreshold='{}'",
     "test:coverage-check": "jest --config './jest.coverage.config.js' --coverage",
@@ -81,6 +81,7 @@
     "eslint": "6.8.0",
     "eslint-config-prettier": "6.0.0",
     "eslint-config-standard": "14.1.1",
+    "eslint-import-resolver-typescript": "^2.2.0",
     "eslint-plugin-cucumber": "^1.4.0",
     "eslint-plugin-import": "2.20.2",
     "eslint-plugin-node": "11.1.0",
@@ -100,6 +101,7 @@
     "standard-version": "^8.0.2",
     "ts-jest": "^26.0.0",
     "ts-node": "8.3.0",
+    "tsconfig-paths": "^3.9.0",
     "typescript": "3.9.2"
   },
   "dependencies": {
