+++ test/step-definitions/template.step.ts
@@ -1,7 +1,7 @@
 import path from 'path'
 import { loadFeature, defineFeature } from 'jest-cucumber'
 import { Server, ServerInjectResponse } from '@hapi/hapi'
-import Config from '../../src/shared/config'
+import Config from '~/shared/config'
 
 import AuthService from '../../src/server'
