+++ test/unit/cli.test.ts
@@ -21,13 +21,13 @@
  * Paweł Marzec <pawel.marzec@modusbox.com>
  --------------
  ******/
-import Config from '../../src/shared/config'
-import server from '../../src/server'
-jest.mock('../../src/server')
+import Config from '~/shared/config'
+import server from '~/server'
+jest.mock('~/server')
 
 describe('cli', (): void => {
   it('should use default port & host', async (): Promise<void> => {
-    const cli = await import('../../src/cli')
+    const cli = await import('~/cli')
     expect(cli).toBeDefined()
     expect(server.run).toHaveBeenCalledWith({
       PACKAGE: Config.PACKAGE,
