+++ test/unit/model/consent.test.ts
@@ -37,8 +37,8 @@
 
 import Knex from 'knex'
 import Config from '../../../config/knexfile'
-import ConsentDB, { Consent } from '../../../src/model/consent'
-import { NotFoundError } from '../../../src/model/errors'
+import ConsentDB, { Consent } from '~/model/consent'
+import { NotFoundError } from '~/model/errors'
 
 /*
  * Mock Consent Resources
