+++ test/unit/model/scope.test.ts
@@ -29,9 +29,9 @@
 
 import Knex from 'knex'
 import Config from '../../../config/knexfile'
-import ScopeDB, { Scope } from '../../../src/model/scope'
-import { Consent } from '../../../src/model/consent'
-import { NotFoundError } from '../../../src/model/errors'
+import ScopeDB, { Scope } from '~/model/scope'
+import { Consent } from '~/model/consent'
+import { NotFoundError } from '~/model/errors'
 
 /*
  * Mock Consent Resources
