+++ test/unit/seeds/consent.test.ts
@@ -24,14 +24,13 @@
  ******/
 
 import Config from '../../../config/knexfile'
-import * as Knex from 'knex'
-import knex from 'knex'
+import Knex from 'knex'
 
 describe('testing Consent table', (): void => {
-  let db: knex<unknown[]>
+  let db: Knex<unknown[]>
 
   beforeAll(async (): Promise<void> => {
-    db = knex(Config.test)
+    db = Knex(Config.test)
     await db.migrate.latest()
     await db.seed.run()
   })
@@ -78,10 +77,10 @@ describe('testing Consent table', (): void => {
 })
 
 describe('testing that constraints are enforced in the consent table', (): void => {
-  let db: knex<unknown[]>
+  let db: Knex<unknown[]>
 
   beforeAll(async (): Promise<void> => {
-    db = knex(Config.test)
+    db = Knex(Config.test)
     await db.migrate.latest()
     await db.seed.run()
   })
