+++ test/unit/seeds/scope.test.ts
@@ -24,14 +24,13 @@
  ******/
 
 import Config from '../../../config/knexfile'
-import knex from 'knex'
-import * as Knex from 'knex'
+import Knex from 'knex'
 
 describe('testing scope table', (): void => {
-  let db: knex<unknown[]>
+  let db: Knex<unknown[]>
 
   beforeAll(async (): Promise<void> => {
-    db = knex(Config.test)
+    db = Knex(Config.test)
     await db.migrate.latest()
     await db.seed.run()
   })
@@ -45,19 +44,19 @@ describe('testing scope table', (): void => {
     const users: Knex.QueryBuilder[] = await db.from('Scope').select('*')
     expect(users.length).toEqual(3)
     expect(users[0]).toEqual({
-      id: 1,
+      id: expect.any(Number),
       consentId: '123',
       action: 'accounts.getBalance',
       accountId: '12345-67890'
     })
     expect(users[1]).toEqual({
-      id: 2,
+      id: expect.any(Number),
       consentId: '123',
       action: 'accounts.transfer',
       accountId: '12345-67890'
     })
     expect(users[2]).toEqual({
-      id: 3,
+      id: expect.any(Number),
       consentId: '124',
       action: 'accounts.transfer',
       accountId: '21345-67890'
@@ -66,10 +65,10 @@ describe('testing scope table', (): void => {
 })
 
 describe('testing that constraints are enforced in the Scope table', (): void => {
-  let db: knex<unknown[]>
+  let db: Knex<unknown[]>
 
   beforeAll(async (): Promise<void> => {
-    db = knex(Config.test)
+    db = Knex(Config.test)
     await db.migrate.latest()
     await db.seed.run()
   })
