+++ test/unit/server/handlers/consents.test.ts
@@ -27,8 +27,8 @@
  --------------
  ******/
 import { Request, ResponseToolkit, ResponseObject } from '@hapi/hapi'
-import { post } from '../../../../src/server/handlers/consents'
-import * as Domain from '../../../../src/domain/consents'
+import { post } from '~/server/handlers/consents'
+import * as Domain from '~/domain/consents'
 import Logger from '@mojaloop/central-services-logger'
 
 const mockStoreConsent = jest.spyOn(Domain, 'createAndStoreConsent')
