+++ tsconfig.json
@@ -1,20 +1,24 @@
 {
   "include": [
     "src",
-    "ambient.d.ts"
+    "test",
+    "ambient.d.ts",
+    "commitlint.config.js",
+    "jest.bdd.config.js",
+    "jest.config.js"
   ],
   "exclude": [
     "node_modules",
-    "**/*.spec.ts",
-    "**/*.test.ts",
-    "test",
-    "lib",
-    "coverage",
-    "**/__mocks__/*"
+    "dist",
+    "coverage"
   ],
   "compilerOptions": {
     "target": "es2018",
     "module": "commonjs",
+    "rootDirs": [
+      "src",
+      "test"
+    ],
     "lib": [
       "esnext"
     ],
@@ -37,9 +41,11 @@
     "moduleResolution": "node",
     "baseUrl": "./",
     "paths": {
-      "*": [
-        "src/*",
-        "node_modules/*"
+      "~/*": [
+        "src/*"
+      ],
+      "test/*": [
+        "test/*"
       ]
     },
     "esModuleInterop": true,
